requirements.txt
fastapi==0.115.6
pydantic==2.10.5
uvicorn==0.34.0
urllib3==2.3.0
crewai==0.95.0
crewai-tools==0.25.8
python-dotenv==1.0.1
mangum
